Organize file structure and add to README
