+++ README.md
@@ -1,1 +1,5 @@
 # deed-map
+
+## HOW TO USE
+
+1. Add your own Google Maps API token in the export.default statement in MapContainer.js as the value for apiKey
