+++ client/main.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App.js';
+import App from '../client/App.js';
 
 ReactDOM.render(<App />, document.getElementById('app'));
