+++ webpack.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-  entry: './main.js',
+  entry: './client/main.js',
   output: {
     path: path.join(__dirname, '/public'),
     filename: 'index.js',
